pyproject.toml
[project]
name = "jdhe-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "pytz>=2026.1.post1",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
